+++ client/src/networking/packet.c
@@ -363,6 +363,7 @@ void Packet_HandleDefineBlock(void) {
     definition.renderType = Packet_ReadByte();
     definition.colliderType = Packet_ReadByte();
     definition.lightType = Packet_ReadByte();
+    definition.lightLevel = Packet_ReadByte();   /* v65.8.1 */
     for (int i = 0; i < 3; i++) definition.min[i] = Packet_ReadByte();
     for (int i = 0; i < 3; i++) definition.max[i] = Packet_ReadByte();
     if (!Block_ApplyDefinition(id, &definition)) {